reformat repo, fix getOrCreates to be generic when possible, loan V2 and loan V3, fix for FDT accounting WIP
