+++ subgraphs/maple-finance/src/mappings/mapleGlobals.ts
@@ -1,14 +1,15 @@
+import { BigDecimal } from "@graphprotocol/graph-ts";
+
 import {
     GlobalsParamSet as GlobalsParamSetEvent,
     SetValidPoolFactoryCall
 } from "../../generated/MapleGlobals/MapleGlobals";
-import { BigDecimal } from "@graphprotocol/graph-ts";
 import { PoolFactory as PoolFactoryTemplate } from "../../generated/templates";
 import { _PoolFactory } from "../../generated/schema";
 
 import { PROTOCOL_GLOBAL_PARAMS_TREASURY_FEE_KEY } from "../common/constants";
-import { getOrCreatePoolFactory } from "../common/mapping_helpers/poolFactory";
-import { getOrCreateProtocol } from "../common/mapping_helpers/protocol";
+import { createEventFromCall } from "../common/utils";
+import { getOrCreatePoolFactory, getOrCreateProtocol } from "../common/mappingHelpers/getOrCreate/spawners";
 
 export function handleGlobalsParamSet(event: GlobalsParamSetEvent): void {
     if (PROTOCOL_GLOBAL_PARAMS_TREASURY_FEE_KEY == event.params.which.toString()) {
@@ -31,5 +32,6 @@ export function handleSetValidPoolFactory(call: SetValidPoolFactoryCall): void {
     getOrCreateProtocol();
 
     // Create pool factory entity
-    getOrCreatePoolFactory(poolFactoryAddress, call.block.timestamp, call.block.number);
+    const eventFromCall = createEventFromCall(call);
+    getOrCreatePoolFactory(eventFromCall, poolFactoryAddress);
 }
