+++ subgraphs/maple-finance/src/mappings/mapleRewardsFactory.ts
@@ -1,7 +1,7 @@
 import { MplRewardsCreated as MplRewardsCreatedEvent } from "../../generated/MapleRewardsFactory/MapleRewardsFactory";
 import { MplReward as MplRewardTemplate } from "../../generated/templates";
 
-import { getOrCreateMplReward } from "../common/mapping_helpers/mplReward";
+import { getOrCreateMplReward } from "../common/mappingHelpers/getOrCreate/markets";
 
 export function handleMplRewardsCreated(event: MplRewardsCreatedEvent): void {
     const mplRewardAddress = event.params.mplRewards;
@@ -10,5 +10,5 @@ export function handleMplRewardsCreated(event: MplRewardsCreatedEvent): void {
     MplRewardTemplate.create(mplRewardAddress);
 
     // Create mpl rewards entity
-    getOrCreateMplReward(mplRewardAddress, event.params.stakingToken, event.params.rewardsToken, event.block.number);
+    getOrCreateMplReward(event, mplRewardAddress);
 }
