+++ subgraphs/maple-finance/subgraph.yaml
@@ -14,7 +14,7 @@ dataSources:
           kind: ethereum/events
           apiVersion: 0.0.5
           language: wasm/assemblyscript
-          file: ./src/mappings/maple-globals.ts
+          file: ./src/mappings/mapleGlobals.ts
           entities:
               - PoolFactory
               - LendingProtocol
@@ -39,7 +39,7 @@ dataSources:
           kind: ethereum/events
           apiVersion: 0.0.5
           language: wasm/assemblyscript
-          file: ./src/mappings/maple-rewards-factory.ts
+          file: ./src/mappings/mapleRewardsFactory.ts
           entities:
               - Market
               - RewardTokens
@@ -56,6 +56,8 @@ dataSources:
                 file: ./abis/ERC20SymbolBytes.json
               - name: MapleGlobals
                 file: ./abis/MapleGlobals.json
+              - name: MplReward
+                file: ./abis/MplRewards.json
           eventHandlers:
               - event: MplRewardsCreated(indexed address,indexed address,indexed address,address)
                 handler: handleMplRewardsCreated
@@ -70,7 +72,7 @@ templates:
           kind: ethereum/events
           apiVersion: 0.0.5
           language: wasm/assemblyscript
-          file: ./src/mappings/pool-factory.ts
+          file: ./src/mappings/poolFactory.ts
           entities:
               - Market
               - _StakeLocker
@@ -84,6 +86,8 @@ templates:
                 file: ./abis/ERC20NameBytes.json
               - name: ERC20SymbolBytes
                 file: ./abis/ERC20SymbolBytes.json
+              - name: StakeLocker
+                file: ./abis/StakeLocker.json
           eventHandlers:
               - event: PoolCreated(indexed address,indexed address,address,address,address,address,uint256,uint256,uint256,string,string)
                 handler: handlePoolCreated
@@ -110,8 +114,10 @@ templates:
           abis:
               - name: Pool
                 file: ./abis/Pool.json
-              - name: Loan
-                file: ./abis/Loan.json
+              - name: LoanV1
+                file: ./abis/LoanV1.json
+              - name: LoanV2OrV3
+                file: ./abis/LoanV2OrV3.json
               - name: ERC20
                 file: ./abis/ERC20.json
               - name: ERC20NameBytes
@@ -126,6 +132,10 @@ templates:
                 file: ./abis/YearnOracle.json
               - name: PoolLib
                 file: ./abis/PoolLib.json
+              - name: StakeLocker
+                file: ./abis/StakeLocker.json
+              - name: MplReward
+                file: ./abis/MplRewards.json
           eventHandlers:
               - event: Transfer(indexed address,indexed address,uint256)
                 handler: handleTransfer
@@ -137,6 +147,8 @@ templates:
                 handler: handleClaim
               - event: DefaultSuffered(indexed address,uint256,uint256,uint256,uint256)
                 handler: handleDefaultSuffered
+              - event: FundsWithdrawn(indexed address,uint256,uint256)
+                handler: handleFundsWithdrawn
 
     - name: StakeLocker
       kind: ethereum/contract
@@ -147,7 +159,7 @@ templates:
           kind: ethereum/events
           apiVersion: 0.0.5
           language: wasm/assemblyscript
-          file: ./src/mappings/stake-locker.ts
+          file: ./src/mappings/stakeLocker.ts
           entities:
               - _StakeLocker
               - _Stake
@@ -172,22 +184,24 @@ templates:
                 file: ./abis/YearnOracle.json
               - name: PoolLib
                 file: ./abis/PoolLib.json
+              - name: MplReward
+                file: ./abis/MplRewards.json
           eventHandlers:
               - event: Stake(indexed address,uint256)
                 handler: handleStake
               - event: Unstake(indexed address,uint256)
                 handler: handleUnstake
 
-    - name: Loan
+    - name: LoanV1
       kind: ethereum/contract
       network: mainnet
       source:
-          abi: Loan
+          abi: LoanV1
       mapping:
           kind: ethereum/events
           apiVersion: 0.0.5
           language: wasm/assemblyscript
-          file: ./src/mappings/loan.ts
+          file: ./src/mappings/loanV1.ts
           entities:
               - _Loan
               - Borrow
@@ -198,8 +212,10 @@ templates:
               - _MplReward
               - Token
           abis:
-              - name: Loan
-                file: ./abis/Loan.json
+              - name: LoanV1
+                file: ./abis/LoanV1.json
+              - name: LoanV2OrV3
+                file: ./abis/LoanV2OrV3.json
               - name: ERC20
                 file: ./abis/ERC20.json
               - name: ERC20NameBytes
@@ -214,6 +230,8 @@ templates:
                 file: ./abis/YearnOracle.json
               - name: PoolLib
                 file: ./abis/PoolLib.json
+              - name: MplReward
+                file: ./abis/MplRewards.json
           eventHandlers:
               - event: Drawdown(uint256)
                 handler: handleDrawdown
@@ -221,6 +239,54 @@ templates:
                 handler: handlePaymentMade
               - event: Liquidation(uint256,uint256,uint256,uint256)
                 handler: handleLiquidation
+
+    - name: LoanV2OrV3
+      kind: ethereum/contract
+      network: mainnet
+      source:
+          abi: LoanV2OrV3
+      mapping:
+          kind: ethereum/events
+          apiVersion: 0.0.5
+          language: wasm/assemblyscript
+          file: ./src/mappings/loanV2OrV3.ts
+          entities:
+              - _Loan
+              - Borrow
+              - Repay
+              - Liquidate
+              - Market
+              - _StakeLocker
+              - _MplReward
+              - Token
+          abis:
+              - name: LoanV1
+                file: ./abis/LoanV1.json
+              - name: LoanV2OrV3
+                file: ./abis/LoanV2OrV3.json
+              - name: ERC20
+                file: ./abis/ERC20.json
+              - name: ERC20NameBytes
+                file: ./abis/ERC20NameBytes.json
+              - name: ERC20SymbolBytes
+                file: ./abis/ERC20SymbolBytes.json
+              - name: MapleGlobals
+                file: ./abis/MapleGlobals.json
+              - name: ChainLinkOracle
+                file: ./abis/ChainLinkOracle.json
+              - name: YearnOracle
+                file: ./abis/YearnOracle.json
+              - name: PoolLib
+                file: ./abis/PoolLib.json
+              - name: MplReward
+                file: ./abis/MplRewards.json
+          eventHandlers:
+              - event: FundsDrawnDown(uint256,indexed address)
+                handler: handleFundsDrawnDown
+              - event: PaymentMade(uint256,uint256)
+                handler: handlePaymentMade
+              - event: Repossessed(uint256,uint256,indexed address)
+                handler: handleRepossessed
 
     - name: MplReward
       kind: ethereum/contract
@@ -231,7 +297,7 @@ templates:
           kind: ethereum/events
           apiVersion: 0.0.5
           language: wasm/assemblyscript
-          file: ./src/mappings/mpl-rewards.ts
+          file: ./src/mappings/mplRewards.ts
           entities:
               - Market
               - _MplReward
@@ -255,6 +321,8 @@ templates:
                 file: ./abis/YearnOracle.json
               - name: PoolLib
                 file: ./abis/PoolLib.json
+              - name: StakeLocker
+                file: ./abis/StakeLocker.json
           eventHandlers:
               - event: Staked(indexed address,uint256)
                 handler: handleStaked
